Rykket rundt på elementer, samt oprettet ny component, og service fil til at holde på interfaces
