+++ src/pages/ProductPage.tsx
@@ -4,26 +4,17 @@ import ProductNavbar from "../components/Product/ProductNavbar";
 import ProductPageCards from "../components/Product/ProductPageCards";
 import { getProducts } from "../service/apiFacade";
 import testCards from "../tests/testProductCardInfo";
-
-// Interface for Product properties
-export interface ProductProps {
-    id: number;
-    name: string;
-    price: number;
-    image: string;
-    qty: number;
-    onSale: boolean;
-    discountPrice: number;
-    category: string;
-}
+import TopImageWithText from "../components/Product/TopImageWithText";
+import { ProductCategory, ProductColor, ProductSizes, ProductMaterial, ProductProps } from "../service/ProductProps.ts";
 
 function ProductPage() {
     // State for products
     const [products, setProducts] = useState<ProductProps[]>([]);
+    console.log(products);
 
     // Test cards for not yet implemented props
     const testCardsImages = testCards.map((card) => card.imgSrc);
-    const testCardsCategory = testCards.map((card) => card.category);
+    // const testCardsCategory = testCards.map((card) => card.category);
     // console.log(testCardsImages);
 
     // Fetch products from API
@@ -31,7 +22,14 @@ function ProductPage() {
         const fetchData = async () => {
             try {
                 const productsData = await getProducts();
-                setProducts(productsData);
+                const transformedProducts = productsData.map((product: ProductProps) => ({
+                    ...product,
+                    categories: product.categories.map((cat: ProductCategory) => cat.name),
+                    colors: product.colors.map((color: ProductColor) => color.colorName),
+                    sizes: product.sizes.map((size: ProductSizes) => size.sizeName),
+                    materials: product.materials.map((material: ProductMaterial) => material.name),
+                }));
+                setProducts(transformedProducts);
             } catch (error) {
                 console.error("Error fetching products:", error);
             }
@@ -41,43 +39,11 @@ function ProductPage() {
 
     return (
         <>
-            <div className="container-fluid">
-                {/* Brug container-fluid for at fjerne begrænsningerne af containeren */}
-
-                {/* Mulig ændring kan laves gennem Image Overlays */}
-                {/* https://getbootstrap.com/docs/5.3/components/card/#image-overlays */}
-                {/* Første sektion med billedet */}
-                <div className="row">
-                    <div className="col position-relative p-0">
-                        {/* Billede strækker sig fra kanten af hver side */}
-                        <img
-                            src="../../img/model-banner-products-001.png"
-                            alt="placeholder"
-                            className="img-fluid overflow position-absolute top-0 start-0 w-100"
-                        />
-                        {/* Tekst på billedet */}
-                        <div className="position-absolute top-0 end-0 p-0 d-none d-md-block">
-                            {/* d-none d-md-block skjuler teksten på mindre skærme */}
-
-                            <div className="row mx-0" style={{ height: "calc(3/8 * 90vw" }}>
-                                <div className="col"></div>
-                                <div className="col my-auto text-end" style={{ marginRight: "8rem" }}>
-                                    <h1>Smukke Smykker</h1>
-                                    <p style={{ maxWidth: "1200px" }}>
-                                        Med en sublim blanding af æstetik og håndværk inviterer vores smykker dig ind i
-                                        en verden af tidsløs elegance og personlig stil. Hvert stykke er nøje udformet
-                                        med en forfinet sans for detaljer og en lidenskab for at skabe unikke
-                                        kunstværker, der udstråler skønhed og karakter.
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
+            {/* Showcase iamge with text component */}
+            <TopImageWithText />
 
             {/* Ny container til næste sektion */}
-            <div className="container" style={{ marginTop: "calc(3/8 * 105vw)" }}>
+            <div className="container" style={{ marginTop: "calc(3/8 * 102vw)" }}>
                 {/* marginTop er sat til 9/21 dele af bredden på billedet, */}
                 {/* således at næste container med content kommer frem under billedet */}
 
@@ -110,9 +76,10 @@ function ProductPage() {
                                     cardQty={product.qty}
                                     cardOnSale={product.onSale}
                                     cardDiscountPrice={product.discountPrice}
-                                    // =============
-                                    // cardCategory={product.category}
-                                    cardCategory={testCardsCategory[index]}
+                                    cardCategory={product.categories}
+                                    cardColors={product.colors}
+                                    cardSizes={product.sizes}
+                                    cardMaterials={product.materials}
                                 />
                             </NavLink>
                         </div>
